int16 to int256 and sign checks in engine

+++ src/ISphereXEngine.sol
@@ -10,8 +10,8 @@ pragma solidity >=0.5.0;
  * @dev Full docs of these functions can be found in SphereXEngine
  */
 interface ISphereXEngine {
-    function sphereXValidatePre(int16 num, address sender, bytes calldata data) external returns (bytes32[] memory);
-    function sphereXValidatePost(int16 num, uint256 gas, bytes32[] calldata valuesBefore, bytes32[] calldata valuesAfter) external;
-    function sphereXValidateInternalPre(int16 num) external;
-    function sphereXValidateInternalPost(int16 num, uint256 gas) external;
+    function sphereXValidatePre(int256 num, address sender, bytes calldata data) external returns (bytes32[] memory);
+    function sphereXValidatePost(int256 num, uint256 gas, bytes32[] calldata valuesBefore, bytes32[] calldata valuesAfter) external;
+    function sphereXValidateInternalPre(int256 num) external;
+    function sphereXValidateInternalPost(int256 num, uint256 gas) external;
 }
